requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Werkzeug==3.1.6
oracledb==3.4.2
tqdm==4.67.3
numpy==2.4.2
opencv-python-headless==4.13.0.92
pillow==12.1.1
torch==2.10.0
torchvision==0.25.0
transformers==5.3.0
paddleocr==3.4.0
paddlepaddle==3.3.0
pytesseract==0.3.13
requests==2.32.5
python-bidi==0.6.7
pyclipper==1.4.0
shapely==2.1.2
pypdfium2==5.5.0
paddlex==3.4.2
prettytable==3.17.0
pyyaml==6.0.2
typing_extensions==4.15.0
python-dateutil==2.9.0.post0
regex==2026.2.28
safetensors==0.7.0
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0
six==1.17.0
certifi==2026.2.25
urllib3==2.6.3
idna==3.11
charset-normalizer==3.4.5
psutil==7.2.2
colorama==0.4.6
markupsafe==3.0.3
prettytable==3.17.0
py-cpuinfo==9.0.0
pycryptodome==3.23.0
pydantic==2.12.5
rich==14.3.3
ujson==5.11.0
regex==2026.2.28
fsspec==2026.2.0
filelock==3.25.0
huggingface_hub==1.6.0
tokenizers==0.22.2
opt-einsum==3.3.0
protobuf==7.34.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1
typing-inspection==0.4.2
tzdata==2025.3
wcwidth==0.6.0
packaging==26.0
shellingham==1.5.4
pygments==2.19.2
mdurl==0.1.2
markdown-it-py==4.0.0
colorlog==6.10.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
bce-python-sdk==0.9.63
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cryptography==46.0.5
huggingface_hub==1.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
 MarkupSafe==3.0.3
packaging==26.0
prettytable==3.17.0
pyyaml==6.0.2
rich==14.3.3
MarkupSafe==3.0.3
safetensors==0.7.0
tokenizers==0.22.2
ujson==5.11.0
urllib3==2.6.3
Werkzeug==3.1.6
